Updating Poor AI and Poor AI 368
