+++ .idea/Checkers_Student.iml
@@ -5,7 +5,7 @@
       <sourceFolder url="file://$MODULE_DIR$/src/checkers-python" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/src/checkers-python/AI_Extensions" isTestSource="false" />
     </content>
-    <orderEntry type="jdk" jdkName="Remote Python 3.6.8 (sftp://tingyind@openlab.ics.uci.edu:22/bin/python3)" jdkType="Python SDK" />
+    <orderEntry type="jdk" jdkName="Python 3.5 (Checkers_Student)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
   <component name="TestRunnerService">
