+++ .idea/deployment.xml
@@ -1,21 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="PublishConfigData" exclude=".svn;.cvs;.idea;.DS_Store;.git;.hg;*.hprof;">
+  <component name="PublishConfigData" autoUpload="On explicit save action" exclude=".svn;.cvs;.idea;.DS_Store;.git;.hg;*.hprof;" autoUploadExternalChanges="true">
     <serverData>
-      <paths name="yutil15@openlab.ics.uci.edu:22">
+      <paths name="tingyind@openlab.ics.uci.edu:22">
         <serverdata>
           <mappings>
-            <mapping deploy="/home/yutil15/Checkers_ShallowBlue" local="$PROJECT_DIR$" web="/" />
-          </mappings>
-        </serverdata>
-      </paths>
-      <paths name="yutil15@openlab.ics.uci.edu:22 (1)">
-        <serverdata>
-          <mappings>
-            <mapping local="$PROJECT_DIR$" web="/" />
+            <mapping deploy="/home/tingyind/cs171/Checkers_ShallowBlue" local="$PROJECT_DIR$" web="/" />
+            <mapping local="" />
           </mappings>
         </serverdata>
       </paths>
     </serverData>
+    <option name="myAutoUpload" value="ON_EXPLICIT_SAVE" />
   </component>
 </project>
